update the unit tests
